+++ src/core/flavors.ts
@@ -1,15 +1,18 @@
 import { readdirSync } from 'fs';
 import path from 'path';
-import { FlavorFunction } from '../typesAndConsts';
+import { Program } from '../main/consts.js';
+import { FlavorFunction } from '../main/typesAndConsts.js';
 
-const flavorsDirPath = path.join(__dirname, '..', 'flavors');
+
+
+const flavorsDirPath = path.join(Program.srcPath, 'flavors');
 
 /** Sync, so it's available from the start. */
 function getAvailableFlavors() {
   let flavors = readdirSync(flavorsDirPath);
   // omit _todo from flavors
-  flavors = flavors.filter(f => !['_todo'].includes(f));
-  flavors = flavors.map(flavor => flavor.replace(/\.[tj]s/, ''));
+  flavors = flavors.filter((f) => !['_todo'].includes(f));
+  flavors = flavors.map((flavor) => flavor.replace(/\.[tj]s/, ''));
   return flavors;
 }
 
@@ -18,5 +21,5 @@ export const availableFlavors = getAvailableFlavors();
 export async function getFlavorFunction(flavor: string): Promise<FlavorFunction> {
   if (!availableFlavors.includes(flavor))
     throw new Error(`Invalid flavor! flavor=${flavor}, availableFlavors=${availableFlavors}`);
-  return (await import(path.join(flavorsDirPath, flavor))).default;
+  return (await import(path.join(flavorsDirPath, `${flavor}.js`))).default;
 }