+++ templates/ts/.eslintrc.js
@@ -12,7 +12,7 @@ module.exports = {
     "ecmaVersion": 12,
     "sourceType": "module",
     "tsconfigRootDir": __dirname,
-    "project": ['./tsconfig.json'],
+    "project": ['./tsconfig.lint.json'],
   },
   "ignorePatterns": [".eslintrc.js"],
   "rules": {